+++ frontend/src/locales/lang/zh-CN/routes/dashboard.ts
@@ -1,5 +1,5 @@
 export default {
-  dashboard: '首页',
+  dashboard: '仪表盘',
   about: '关于',
   workbench: '工作台',
   analysis: '分析页',